Ref: Renamed card_bottom/top_details
