+++ app/src/components/common/index.ts
@@ -24,7 +24,7 @@ export { Select } from './form/select'
 export { Spinner } from '../loading/spinner'
 export { SubsectionTitle } from './text/subsection_title'
 export { SubsectionTitleAction } from './text/subsection_title_action'
-export { SubsectionTitleWrapper } from '../market/sections/common_sections/top_card_details/subsection_title_wrapper'
+export { SubsectionTitleWrapper } from '../market/sections/common_sections/card_top_details/subsection_title_wrapper'
 export { TD, TH, THead, TR, Table } from '../market/sections/common_sections/tables/table'
 export { Textfield } from './form/textfield'
 export { SimpleTextfield } from './form/simiple_textfield'
