+++ app/src/components/market/sections/common_sections/card_bottom_details/categorical/bar_diagram_probabilities/index.tsx
