+++ app/src/components/market/sections/common_sections/card_bottom_details/grid_transaction_details/index.tsx
