+++ app/src/components/market/sections/common_sections/card_bottom_details/scalar/market_scale/value_box.tsx
