+++ app/src/components/market/sections/common_sections/card_top_details/additional_market_data/index.tsx
