+++ app/src/components/market/sections/common_sections/card_top_details/market_data/index.tsx
