+++ app/src/components/market/sections/common_sections/card_top_details/market_top_details_open/index.tsx
