+++ app/src/components/market/sections/market_buy/scalar_market_buy.tsx
@@ -45,8 +45,8 @@ import { BigNumberInputReturn } from '../../../common/form/big_number_input'
 import { ModalTransactionWrapper } from '../../../modal'
 import { CurrenciesWrapper, GenericError, TabsGrid } from '../../common/common_styled'
 import { SetAllowance } from '../common_sections/allowance/set_allowance'
-import { GridTransactionDetails } from '../common_sections/bottom_card_details/grid_transaction_details'
-import { MarketScale } from '../common_sections/bottom_card_details/scalar/market_scale'
+import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
+import { MarketScale } from '../common_sections/card_bottom_details/scalar/market_scale'
 import { WarningMessage } from '../common_sections/messages/warning_message'
 import { CurrencySelector } from '../common_sections/user_transactions_tokens/currency_selector'
 import { TransactionDetailsCard } from '../common_sections/user_transactions_tokens/transaction/transaction_details_card'
