+++ app/src/components/market/sections/market_create/steps/items/funding_and_fee_step.tsx
@@ -44,8 +44,8 @@ import { CreateCard } from '../../../../common/create_card'
 import { VerifiedRow } from '../../../../common/verified_row'
 import { SetAllowance } from '../../../common_sections/allowance/set_allowance'
 import { DisplayArbitrator } from '../../../common_sections/arbitrators/display_arbitrator'
-import { GridTransactionDetails } from '../../../common_sections/bottom_card_details/grid_transaction_details'
-import { MarketScale } from '../../../common_sections/bottom_card_details/scalar/market_scale'
+import { GridTransactionDetails } from '../../../common_sections/card_bottom_details/grid_transaction_details'
+import { MarketScale } from '../../../common_sections/card_bottom_details/scalar/market_scale'
 import { WarningMessage } from '../../../common_sections/messages/warning_message'
 import { CurrencySelector } from '../../../common_sections/user_transactions_tokens/currency_selector'
 import { TradingFeeSelector } from '../../../common_sections/user_transactions_tokens/trading_fee_selector'
