+++ app/src/components/market/sections/market_pooling/market_pool_liquidity.tsx
@@ -51,7 +51,7 @@ import { ModalTransactionWrapper } from '../../../modal'
 import { CurrenciesWrapper, GenericError, TabsGrid } from '../../common/common_styled'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
 import { SetAllowance } from '../common_sections/allowance/set_allowance'
-import { GridTransactionDetails } from '../common_sections/bottom_card_details/grid_transaction_details'
+import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
 import { WarningMessage } from '../common_sections/messages/warning_message'
 import { OutcomeTable } from '../common_sections/tables/outcome_table'
 import { CurrencySelector } from '../common_sections/user_transactions_tokens/currency_selector'
