+++ app/src/components/market/sections/market_pooling/scalar_market_pool_liquidity.tsx
@@ -54,8 +54,8 @@ import { ModalTransactionWrapper } from '../../../modal'
 import { CurrenciesWrapper, GenericError, TabsGrid } from '../../common/common_styled'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
 import { SetAllowance } from '../common_sections/allowance/set_allowance'
-import { GridTransactionDetails } from '../common_sections/bottom_card_details/grid_transaction_details'
-import { MarketScale } from '../common_sections/bottom_card_details/scalar/market_scale'
+import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
+import { MarketScale } from '../common_sections/card_bottom_details/scalar/market_scale'
 import { WarningMessage } from '../common_sections/messages/warning_message'
 import { CurrencySelector } from '../common_sections/user_transactions_tokens/currency_selector'
 import { TransactionDetailsCard } from '../common_sections/user_transactions_tokens/transaction/transaction_details_card'
