+++ app/src/components/market/sections/market_profile/market_status/closed.tsx
@@ -27,10 +27,10 @@ import { ButtonType } from '../../../../button/button_styling_types'
 import { ModalTransactionWrapper } from '../../../../modal'
 import { MarginsButton } from '../../../common/common_styled'
 import { ViewCard } from '../../../common/view_card'
-import { MarketScale } from '../../common_sections/bottom_card_details/scalar/market_scale'
+import { MarketScale } from '../../common_sections/card_bottom_details/scalar/market_scale'
+import { MarketTopDetailsClosed } from '../../common_sections/card_top_details/market_top_details_closed'
 import MarketResolutionMessage from '../../common_sections/messages/market_resolution_message'
 import { OutcomeTable } from '../../common_sections/tables/outcome_table'
-import { MarketTopDetailsClosed } from '../../common_sections/top_card_details/market_top_details_closed'
 import { MarketBuyContainer } from '../../market_buy/market_buy_container'
 import { MarketHistoryContainer } from '../../market_history/market_history_container'
 import { MarketNavigation } from '../../market_navigation'
