+++ app/src/components/market/sections/market_profile/market_status/open.tsx
@@ -19,10 +19,10 @@ import {
 import { Button, ButtonContainer } from '../../../../button'
 import { ButtonType } from '../../../../button/button_styling_types'
 import { ViewCard } from '../../../common/view_card'
-import { MarketScale } from '../../common_sections/bottom_card_details/scalar/market_scale'
+import { MarketScale } from '../../common_sections/card_bottom_details/scalar/market_scale'
+import { MarketTopDetailsOpen } from '../../common_sections/card_top_details/market_top_details_open'
 import { WarningMessage } from '../../common_sections/messages/warning_message'
 import { OutcomeTable } from '../../common_sections/tables/outcome_table'
-import { MarketTopDetailsOpen } from '../../common_sections/top_card_details/market_top_details_open'
 import { MarketBondContainer } from '../../market_bond/market_bond_container'
 import { MarketBuyContainer } from '../../market_buy/market_buy_container'
 import { MarketHistoryContainer } from '../../market_history/market_history_container'
