+++ app/src/components/market/sections/market_sell/market_sell.tsx
@@ -42,7 +42,7 @@ import { BigNumberInput, TextfieldCustomPlaceholder } from '../../../common'
 import { ModalTransactionWrapper } from '../../../modal'
 import { GenericError } from '../../common/common_styled'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
-import { GridTransactionDetails } from '../common_sections/bottom_card_details/grid_transaction_details'
+import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
 import { WarningMessage } from '../common_sections/messages/warning_message'
 import { OutcomeTable } from '../common_sections/tables/outcome_table'
 import { TransactionDetailsCard } from '../common_sections/user_transactions_tokens/transaction/transaction_details_card'
