+++ app/src/components/market/sections/market_sell/scalar_market_sell.tsx
@@ -43,8 +43,8 @@ import { ModalTransactionWrapper } from '../../../modal'
 import { GenericError } from '../../common/common_styled'
 import { PositionSelectionBox } from '../../common/position_selection_box'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
-import { GridTransactionDetails } from '../common_sections/bottom_card_details/grid_transaction_details'
-import { MarketScale } from '../common_sections/bottom_card_details/scalar/market_scale'
+import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
+import { MarketScale } from '../common_sections/card_bottom_details/scalar/market_scale'
 import { WarningMessage } from '../common_sections/messages/warning_message'
 import { TransactionDetailsCard } from '../common_sections/user_transactions_tokens/transaction/transaction_details_card'
 import { TransactionDetailsLine } from '../common_sections/user_transactions_tokens/transaction/transaction_details_line'
